refactor linked hash map into a new file
